% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{allowed_admin_levels}
\alias{allowed_admin_levels}
\title{Allowed GADM admin levels}
\format{
An object of class \code{character} of length 4.
}
\usage{
allowed_admin_levels
}
\description{
Allowed GADM admin levels
}
\keyword{constants}
\keyword{internal}
