% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{data_source_names}
\alias{data_source_names}
\title{Climate data source names}
\format{
An object of class \code{character} of length 7.
}
\usage{
data_source_names
}
\description{
Climate data source names
}
\keyword{constants}
\keyword{internal}
