% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{input_validators}
\alias{input_validators}
\alias{validate_country}
\alias{validate_date_range}
\alias{validate_admin_level}
\title{Return checked inputs or print informative errors}
\usage{
validate_country(country)

validate_date_range(date_range)

validate_admin_level(admin_level)
}
\arguments{
\item{country}{A string for the country name, or for the country ISO 2- or 3-
character code. E.g. one of "Canada", "CA", or "CAN". Input is checked
against names and codes in \pkg{countrycode}.}

\item{date_range}{A two element vector of \verb{<Date>}s, giving the interval for
which to get data.}

\item{admin_level}{A numeric giving the GADM admin level for which to get
data. Only one unit may be passed at a time.}
}
\value{
A checked and in some cases processed input. Most functions also have
side-effects.
\itemize{
\item \code{validate_country()}: A country ISO 3 character code. Has the side-effect
of erroring when \code{country} is either not a string, or when the country code
cannot be found from \code{country}.
\item \code{validate_date_range()}: Returns \code{date_range} after checking that it is
a two-element \verb{<Date>} vector.
\item \code{validate_admin_level()}: Returns \code{admin_level} after checking it is an
integer-like number in the range \verb{[0, 3]}.
}
}
\description{
Return checked inputs or print informative errors
}
\keyword{internal}
\keyword{validator}
