% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vimc_climate.R
\name{load_local_vimc_climate}
\alias{load_local_vimc_climate}
\title{Get climate data prepared for VIMC and stored locally}
\usage{
load_local_vimc_climate(
  country,
  date_range,
  data_location = here::here(),
  data_source = c("CHIRPS", "ERA5_mean", "ERA5_min", "ERA5_max", "ERA5_RH", "ERA5_SH",
    "PERSIANN"),
  admin_level = c(0, 1, 2, 3)
)
}
\arguments{
\item{country}{A string for the country name, or for the country ISO 2- or 3-
character code. E.g. one of "Canada", "CA", or "CAN". Input is checked
against names and codes in \pkg{countrycode}.}

\item{date_range}{A two element vector of \verb{<Date>}s, giving the interval for
which to get data.}

\item{data_location}{A location relative to the current working directory
to search for the data. Data for \code{country} are expected to live in
\code{data_location/XYZ} as Parquet files, where \code{XYZ} is the ISO 3-character code
for \code{country}. Defaults to the working directory.}

\item{data_source}{A string giving the climate data to search for. Names must
match exactly.}

\item{admin_level}{A numeric giving the GADM admin level for which to get
data. Only one unit may be passed at a time.}
}
\value{
A \verb{<data.frame>} of climate data.
}
\description{
Load climate data prepared for VIMC from a local directory.
}
\keyword{data_access}
