% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_data.R
\name{locate_parquet_file}
\alias{locate_parquet_file}
\title{Locate local data files}
\usage{
locate_parquet_file(data_location, country_iso3c, admin_level, data_source)
}
\arguments{
\item{data_location}{A location relative to the current working directory
to search for the data. Data for \code{country} are expected to live in
\code{data_location/XYZ} as Parquet files, where \code{XYZ} is the ISO 3-character code
for \code{country}. Defaults to the working directory.}

\item{country_iso3c}{Country ISO 3 character code. Passed from top-level
function \code{\link[=load_local_vimc_climate]{load_local_vimc_climate()}} after a call to \code{\link[=validate_country]{validate_country()}}.}

\item{admin_level}{A numeric giving the GADM admin level for which to get
data. Only one unit may be passed at a time.}

\item{data_source}{A string giving the climate data to search for. Names must
match exactly.}
}
\value{
The location of a local data Parquet file. Has a number of
side-effects to inform users when \code{data_location}, the county data directory
within \code{data_location}, or the file corresponding to the data requested are
not found.
}
\description{
Locate local data files
}
\keyword{data_locator}
\keyword{internal}
