% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimclimate.R
\docType{package}
\name{vimclimate-package}
\alias{vimclimate}
\alias{vimclimate-package}
\title{vimclimate: Access Climate Data Time-series from the Vaccine Impact Modelling Consortium}
\description{
Access pre-made climate data time-series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vimc/vimclimate}
  \item \url{https://vimc.github.io/vimclimate}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

}
\keyword{internal}
